requirements.txt
fastapi
starlette
psycopg2
plotly
uvicorn
sqlalchemy
psycopg2-binary
sqlalchemy
python-dotenv
jinja2
pydantic
sqlacodegen
passlib[bcrypt]
python-jose
pydantic[email]
dateutil.relativedelta